Move traverser to non recursive
